update categs,metLD,double ev_id, training & loose categ, ...

+++ NtupleProducer/src/ElectronExt.cxx
@@ -260,7 +260,7 @@ void ElectronExt::sel()
 
    for(int d=0;d<evdebug->size();d++)
      {		       
-	int evId = ntP->ev_id;
+	double evId = ntP->ev_id;
 	if( evId == evdebug->at(d) )
 	  {
 	     std::cout << "------------------------------" << std::endl;
